+++ core/src/main/java/com/refinedmods/refinedstorage2/core/network/node/diskdrive/DiskDriveListener.java
@@ -1,0 +1,5 @@
+package com.refinedmods.refinedstorage2.core.network.node.diskdrive;
+
+public interface DiskDriveListener {
+    void onDiskChanged();
+}