+++ core/src/test/java/com/refinedmods/refinedstorage2/core/network/component/EnergyNetworkComponentTest.java
@@ -1,7 +1,6 @@
 package com.refinedmods.refinedstorage2.core.network.component;
 
 import com.refinedmods.refinedstorage2.core.Rs2Test;
-import com.refinedmods.refinedstorage2.core.adapter.FakeRs2World;
 import com.refinedmods.refinedstorage2.core.network.node.container.FakeNetworkNodeContainer;
 import com.refinedmods.refinedstorage2.core.network.node.container.NetworkNodeContainer;
 import com.refinedmods.refinedstorage2.core.network.node.controller.ControllerNetworkNode;
@@ -25,14 +24,12 @@ class EnergyNetworkComponentTest {
     }
 
     private NetworkNodeContainer<ControllerNetworkNode> createControllerContainer() {
-        NetworkNodeContainer<ControllerNetworkNode> container = new FakeNetworkNodeContainer<>(new ControllerNetworkNode(
+        return FakeNetworkNodeContainer.createForFakeWorld(new ControllerNetworkNode(
                 Position.ORIGIN,
                 100,
                 1000,
                 ControllerType.NORMAL
         ));
-        container.setContainerWorld(new FakeRs2World());
-        return container;
     }
 
     @Test