+++ core/src/test/java/com/refinedmods/refinedstorage2/core/network/node/container/FakeNetworkNodeContainer.java
@@ -1,6 +1,7 @@
 package com.refinedmods.refinedstorage2.core.network.node.container;
 
 import com.refinedmods.refinedstorage2.core.Rs2World;
+import com.refinedmods.refinedstorage2.core.adapter.FakeRs2World;
 import com.refinedmods.refinedstorage2.core.network.component.NetworkComponentRegistry;
 import com.refinedmods.refinedstorage2.core.network.node.NetworkNode;
 import com.refinedmods.refinedstorage2.core.util.Direction;
@@ -10,19 +11,27 @@ import java.util.List;
 
 public class FakeNetworkNodeContainer<T extends NetworkNode> implements NetworkNodeContainer<T> {
     private final T node;
+    private final Rs2World world;
 
-    public FakeNetworkNodeContainer(T node) {
+    private FakeNetworkNodeContainer(T node, Rs2World world) {
         this.node = node;
+        this.world = world;
+    }
+
+    public static <T extends NetworkNode> FakeNetworkNodeContainer<T> createForFakeWorld(T node) {
+        Rs2World world = new FakeRs2World();
+        node.setWorld(world);
+        return new FakeNetworkNodeContainer<>(node, world);
     }
 
     @Override
     public boolean initialize(NetworkNodeContainerRepository containerRepository, NetworkComponentRegistry networkComponentRegistry) {
-        return false;
+        throw new UnsupportedOperationException();
     }
 
     @Override
     public void remove(NetworkNodeContainerRepository containerRepository, NetworkComponentRegistry networkComponentRegistry) {
-
+        throw new UnsupportedOperationException();
     }
 
     @Override
@@ -32,26 +41,26 @@ public class FakeNetworkNodeContainer<T extends NetworkNode> implements NetworkN
 
     @Override
     public Position getPosition() {
-        return null;
+        return Position.ORIGIN;
     }
 
     @Override
     public Rs2World getContainerWorld() {
-        return null;
+        return world;
     }
 
     @Override
     public void setContainerWorld(Rs2World world) {
-        node.setWorld(world);
+        throw new UnsupportedOperationException();
     }
 
     @Override
     public List<NetworkNodeContainer<?>> getConnections(NetworkNodeContainerRepository containerRepository) {
-        return null;
+        throw new UnsupportedOperationException();
     }
 
     @Override
     public boolean canConnectWith(NetworkNodeContainer<?> other, Direction incomingDirection) {
-        return false;
+        throw new UnsupportedOperationException();
     }
 }