+++ core/src/test/java/com/refinedmods/refinedstorage2/core/network/node/diskdrive/DiskDriveNetworkNodeTest.java
@@ -5,7 +5,6 @@ import com.refinedmods.refinedstorage2.core.item.ItemStubs;
 import com.refinedmods.refinedstorage2.core.item.Rs2ItemStack;
 import com.refinedmods.refinedstorage2.core.network.Network;
 import com.refinedmods.refinedstorage2.core.network.NetworkUtil;
-import com.refinedmods.refinedstorage2.core.network.component.ItemStorageNetworkComponent;
 import com.refinedmods.refinedstorage2.core.network.node.RedstoneMode;
 import com.refinedmods.refinedstorage2.core.network.node.container.FakeNetworkNodeContainer;
 import com.refinedmods.refinedstorage2.core.storage.AccessMode;
@@ -14,6 +13,7 @@ import com.refinedmods.refinedstorage2.core.storage.disk.ItemDiskStorage;
 import com.refinedmods.refinedstorage2.core.storage.disk.StorageDisk;
 import com.refinedmods.refinedstorage2.core.util.Action;
 import com.refinedmods.refinedstorage2.core.util.FilterMode;
+import com.refinedmods.refinedstorage2.core.util.Position;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -24,10 +24,18 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.params.provider.ValueSource;
+import org.mockito.verification.VerificationMode;
 
+import static com.refinedmods.refinedstorage2.core.network.NetworkUtil.itemStorageChannelOf;
+import static com.refinedmods.refinedstorage2.core.network.NetworkUtil.itemStorageOf;
 import static com.refinedmods.refinedstorage2.core.util.ItemStackAssertions.assertItemStack;
 import static com.refinedmods.refinedstorage2.core.util.ItemStackAssertions.assertItemStackListContents;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @Rs2Test
 class DiskDriveNetworkNodeTest {
@@ -35,169 +43,150 @@ class DiskDriveNetworkNodeTest {
     private static final long USAGE_PER_DISK = 3;
 
     private Network network;
-    private DiskDriveNetworkNodeWrapper diskDrive;
+    private DiskDriveNetworkNode diskDrive;
+    private FakeStorageDiskProviderManager storageDiskProviderManager;
+    private DiskDriveListener diskDriveListener;
 
     @BeforeEach
     void setUp() {
+        diskDriveListener = mock(DiskDriveListener.class);
+        storageDiskProviderManager = new FakeStorageDiskProviderManager();
+
         network = NetworkUtil.createWithCreativeEnergySource();
-        diskDrive = DiskDriveNetworkNodeWrapper.create(BASE_USAGE, USAGE_PER_DISK);
+
+        diskDrive = createDiskDriveContainer(network, storageDiskProviderManager, diskDriveListener).getNode();
+    }
+
+    private FakeNetworkNodeContainer<DiskDriveNetworkNode> createDiskDriveContainer(Network network, FakeStorageDiskProviderManager storageDiskProviderManager, DiskDriveListener diskDriveListener) {
+        DiskDriveNetworkNode diskDrive = new DiskDriveNetworkNode(Position.ORIGIN, storageDiskProviderManager, BASE_USAGE, USAGE_PER_DISK, diskDriveListener);
         diskDrive.setNetwork(network);
-        network.addContainer(new FakeNetworkNodeContainer<>(diskDrive));
+
+        FakeNetworkNodeContainer<DiskDriveNetworkNode> container = FakeNetworkNodeContainer.createForFakeWorld(diskDrive);
+
+        network.addContainer(container);
+
+        return container;
     }
 
     @Test
-    void Test_initial_disk_state() {
+    void Test_initial_state() {
+        // Arrange
+        ItemDiskStorage disk = new ItemDiskStorage(10);
+        disk.insert(new Rs2ItemStack(ItemStubs.DIRT), 5, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(1, disk);
+
         // Act
         DiskDriveState states = diskDrive.createState();
 
         // Assert
+        assertThat(diskDrive.getEnergyUsage()).isEqualTo(BASE_USAGE);
+        assertThat(diskDrive.getStacks()).isEmpty();
+        assertThat(diskDrive.getStored()).isZero();
         assertThat(states.getStates())
                 .hasSize(DiskDriveNetworkNode.DISK_COUNT)
                 .allMatch(state -> state == DiskState.NONE);
+        assertItemStackListContents(itemStorageChannelOf(network).getStacks());
     }
 
     @Test
-    void Test_state_with_an_empty_disk() {
+    void Test_initialization() {
         // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 74, Action.EXECUTE);
-
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk);
+        ItemDiskStorage disk = new ItemDiskStorage(10);
+        disk.insert(new Rs2ItemStack(ItemStubs.DIRT), 5, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(1, disk);
 
         // Act
-        DiskDriveState state = diskDrive.createState();
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Assert
-        assertThat(state.getState(0)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(1)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(2)).isEqualTo(DiskState.NORMAL);
-        assertThat(state.getState(3)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(4)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(5)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(6)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(7)).isEqualTo(DiskState.NONE);
+        assertThat(diskDrive.getEnergyUsage()).isEqualTo(BASE_USAGE + USAGE_PER_DISK);
+        assertItemStackListContents(diskDrive.getStacks(), new Rs2ItemStack(ItemStubs.DIRT, 5));
+        assertItemStackListContents(itemStorageChannelOf(network).getStacks());
+        assertThat(diskDrive.getStored()).isEqualTo(5L);
     }
 
-    @Test
-    void Test_state_with_a_full_disk() {
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    void Test_disk_state(boolean inactive) {
         // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(10);
-        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 10, Action.EXECUTE);
+        StorageDisk<Rs2ItemStack> normalDisk = new ItemDiskStorage(100);
+        normalDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 74, Action.EXECUTE);
 
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(7, storageDisk);
+        StorageDisk<Rs2ItemStack> nearCapacityDisk = new ItemDiskStorage(100);
+        nearCapacityDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 75, Action.EXECUTE);
+
+        StorageDisk<Rs2ItemStack> fullDisk = new ItemDiskStorage(100);
+        fullDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 100, Action.EXECUTE);
+
+        storageDiskProviderManager.setDiskInSlot(1, UUID.randomUUID());
+        storageDiskProviderManager.setDiskInSlot(3, normalDisk);
+        storageDiskProviderManager.setDiskInSlot(5, nearCapacityDisk);
+        storageDiskProviderManager.setDiskInSlot(7, fullDisk);
+
+        if (inactive) {
+            diskDrive.setRedstoneMode(RedstoneMode.HIGH);
+        }
 
         // Act
-        DiskDriveState state = diskDrive.createState();
+        diskDrive.initialize(storageDiskProviderManager);
 
-        // Assert
-        assertThat(state.getState(0)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(1)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(2)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(3)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(4)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(5)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(6)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(7)).isEqualTo(DiskState.FULL);
-    }
-
-    @Test
-    void Test_state_with_a_nearly_full_disk() {
-        // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 75, Action.EXECUTE);
-
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(3, storageDisk);
-
-        // Act
         DiskDriveState state = diskDrive.createState();
 
         // Assert
         assertThat(state.getState(0)).isEqualTo(DiskState.NONE);
         assertThat(state.getState(1)).isEqualTo(DiskState.NONE);
         assertThat(state.getState(2)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(3)).isEqualTo(DiskState.NEAR_CAPACITY);
+        assertThat(state.getState(3)).isEqualTo(inactive ? DiskState.DISCONNECTED : DiskState.NORMAL);
         assertThat(state.getState(4)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(5)).isEqualTo(DiskState.NONE);
+        assertThat(state.getState(5)).isEqualTo(inactive ? DiskState.DISCONNECTED : DiskState.NEAR_CAPACITY);
         assertThat(state.getState(6)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(7)).isEqualTo(DiskState.NONE);
+        assertThat(state.getState(7)).isEqualTo(inactive ? DiskState.DISCONNECTED : DiskState.FULL);
+        assertThat(diskDrive.getEnergyUsage()).isEqualTo(BASE_USAGE + (USAGE_PER_DISK * 3));
     }
 
-
     @Test
-    void Test_state_when_inactive() {
+    void Test_setting_disk_in_slot() {
         // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 75, Action.EXECUTE);
+        diskDrive.initialize(storageDiskProviderManager);
 
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(3, storageDisk);
-        diskDrive.setRedstoneMode(RedstoneMode.HIGH);
+        ItemDiskStorage disk = new ItemDiskStorage(10);
+        disk.insert(new Rs2ItemStack(ItemStubs.DIRT), 5, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(7, disk);
 
         // Act
-        DiskDriveState state = diskDrive.createState();
+        diskDrive.onDiskChanged(7);
 
         // Assert
-        assertThat(state.getState(0)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(1)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(2)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(3)).isEqualTo(DiskState.DISCONNECTED);
-        assertThat(state.getState(4)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(5)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(6)).isEqualTo(DiskState.NONE);
-        assertThat(state.getState(7)).isEqualTo(DiskState.NONE);
+        assertThat(diskDrive.getEnergyUsage()).isEqualTo(BASE_USAGE + USAGE_PER_DISK);
+        assertItemStackListContents(diskDrive.getStacks(), new Rs2ItemStack(ItemStubs.DIRT, 5));
+        assertItemStackListContents(itemStorageChannelOf(network).getStacks(), new Rs2ItemStack(ItemStubs.DIRT, 5));
+        assertThat(diskDrive.getStored()).isEqualTo(5L);
     }
 
     @Test
-    void Test_state_should_change_when_removing_disk() {
+    void Test_removing_disk_in_slot() {
         // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 74, Action.EXECUTE);
+        diskDrive.initialize(storageDiskProviderManager);
+
+        ItemDiskStorage disk = new ItemDiskStorage(10);
+        disk.insert(new Rs2ItemStack(ItemStubs.DIRT), 5, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(7, disk);
+
+        diskDrive.onDiskChanged(7);
 
         // Act
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(4, storageDisk);
-        DiskDriveState stateBeforeRemoval = diskDrive.createState();
-
-        diskDrive.getFakeStorageDiskProviderManager().removeDiskInSlot(2);
-        DiskDriveState stateAfterRemoval = diskDrive.createState();
+        storageDiskProviderManager.removeDiskInSlot(7);
+        diskDrive.onDiskChanged(7);
 
         // Assert
-        assertThat(stateBeforeRemoval.getState(0)).isEqualTo(DiskState.NONE);
-        assertThat(stateBeforeRemoval.getState(1)).isEqualTo(DiskState.NONE);
-        assertThat(stateBeforeRemoval.getState(2)).isEqualTo(DiskState.NORMAL);
-        assertThat(stateBeforeRemoval.getState(3)).isEqualTo(DiskState.NONE);
-        assertThat(stateBeforeRemoval.getState(4)).isEqualTo(DiskState.NORMAL);
-        assertThat(stateBeforeRemoval.getState(5)).isEqualTo(DiskState.NONE);
-        assertThat(stateBeforeRemoval.getState(6)).isEqualTo(DiskState.NONE);
-        assertThat(stateBeforeRemoval.getState(7)).isEqualTo(DiskState.NONE);
-
-        assertThat(stateAfterRemoval.getState(0)).isEqualTo(DiskState.NONE);
-        assertThat(stateAfterRemoval.getState(1)).isEqualTo(DiskState.NONE);
-        assertThat(stateAfterRemoval.getState(2)).isEqualTo(DiskState.NONE);
-        assertThat(stateAfterRemoval.getState(3)).isEqualTo(DiskState.NONE);
-        assertThat(stateAfterRemoval.getState(4)).isEqualTo(DiskState.NORMAL);
-        assertThat(stateAfterRemoval.getState(5)).isEqualTo(DiskState.NONE);
-        assertThat(stateAfterRemoval.getState(6)).isEqualTo(DiskState.NONE);
-        assertThat(stateAfterRemoval.getState(7)).isEqualTo(DiskState.NONE);
+        assertThat(diskDrive.getEnergyUsage()).isEqualTo(BASE_USAGE);
+        assertItemStackListContents(diskDrive.getStacks());
+        assertItemStackListContents(itemStorageChannelOf(network).getStacks());
+        assertThat(diskDrive.getStored()).isZero();
     }
 
     @Test
-    void Test_state_with_a_disk_that_does_not_exist() {
-        // Arrange
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(3, UUID.randomUUID());
-
-        // Act
-        diskDrive.onDiskChanged(3);
-
-        DiskDriveState states = diskDrive.createState();
-
-        // Assert
-        assertThat(states.getStates())
-                .hasSize(DiskDriveNetworkNode.DISK_COUNT)
-                .allMatch(state -> state == DiskState.NONE);
-    }
-
-    @Test
-    void Test_state_when_changing_an_invalid_slot() {
+    void Test_changing_disk_in_invalid_slot() {
         // Act
         diskDrive.onDiskChanged(-1);
         diskDrive.onDiskChanged(DiskDriveNetworkNode.DISK_COUNT);
@@ -211,21 +200,72 @@ class DiskDriveNetworkNodeTest {
     }
 
     @Test
+    void Test_retrieving_stacks() {
+        // Arrange
+        diskDrive.initialize(storageDiskProviderManager);
+
+        StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
+        storageDisk1.insert(new Rs2ItemStack(ItemStubs.DIRT), 50, Action.EXECUTE);
+        storageDisk1.insert(new Rs2ItemStack(ItemStubs.GLASS), 50, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk1);
+
+        diskDrive.onDiskChanged(1);
+
+        // Act
+        Collection<Rs2ItemStack> stacks = diskDrive.getStacks();
+        Collection<Rs2ItemStack> stacksInNetwork = itemStorageChannelOf(network).getStacks();
+        long storedInNetwork = itemStorageChannelOf(network).getStored();
+
+        // Assert
+        assertItemStackListContents(stacks, new Rs2ItemStack(ItemStubs.DIRT, 50), new Rs2ItemStack(ItemStubs.GLASS, 50));
+        assertItemStackListContents(stacksInNetwork, new Rs2ItemStack(ItemStubs.DIRT, 50), new Rs2ItemStack(ItemStubs.GLASS, 50));
+        assertThat(diskDrive.getStored()).isEqualTo(100);
+        assertThat(storedInNetwork).isEqualTo(100);
+    }
+
+    @Test
+    void Test_retrieving_stacks_when_inactive() {
+        // Arrange
+        diskDrive.initialize(storageDiskProviderManager);
+
+        StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
+        storageDisk1.insert(new Rs2ItemStack(ItemStubs.DIRT), 50, Action.EXECUTE);
+        storageDisk1.insert(new Rs2ItemStack(ItemStubs.GLASS), 50, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk1);
+
+        diskDrive.setRedstoneMode(RedstoneMode.HIGH);
+        diskDrive.onDiskChanged(1);
+
+        // Act
+        Collection<Rs2ItemStack> stacks = diskDrive.getStacks();
+        Collection<Rs2ItemStack> stacksInNetwork = itemStorageChannelOf(network).getStacks();
+        long storedInNetwork = itemStorageChannelOf(network).getStored();
+
+        // Assert
+        assertItemStackListContents(stacks);
+        assertItemStackListContents(stacksInNetwork);
+        assertThat(diskDrive.getStored()).isEqualTo(100L);
+        assertThat(storedInNetwork).isEqualTo(100L);
+    }
+
+    @Test
     void Test_inserting() {
         // Arrange
         StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk1);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk1);
 
         StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk2);
+        storageDiskProviderManager.setDiskInSlot(2, storageDisk2);
 
         StorageDisk<Rs2ItemStack> storageDisk3 = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(3, storageDisk3);
+        storageDiskProviderManager.setDiskInSlot(3, storageDisk3);
+
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Act
-        Optional<Rs2ItemStack> remainder1 = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 150, Action.EXECUTE);
-        Optional<Rs2ItemStack> remainder2 = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 10, Action.EXECUTE);
-        Optional<Rs2ItemStack> remainder3 = diskDrive.insert(new Rs2ItemStack(ItemStubs.GLASS), 300, Action.EXECUTE);
+        Optional<Rs2ItemStack> remainder1 = itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.DIRT), 150, Action.EXECUTE);
+        Optional<Rs2ItemStack> remainder2 = itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.DIRT), 10, Action.EXECUTE);
+        Optional<Rs2ItemStack> remainder3 = itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.GLASS), 300, Action.EXECUTE);
 
         // Assert
         assertThat(remainder1).isEmpty();
@@ -236,6 +276,7 @@ class DiskDriveNetworkNodeTest {
         assertItemStackListContents(storageDisk1.getStacks(), new Rs2ItemStack(ItemStubs.DIRT, 100));
         assertItemStackListContents(storageDisk2.getStacks(), new Rs2ItemStack(ItemStubs.DIRT, 60), new Rs2ItemStack(ItemStubs.GLASS, 40));
         assertItemStackListContents(storageDisk3.getStacks(), new Rs2ItemStack(ItemStubs.GLASS, 100));
+        assertItemStackListContents(itemStorageChannelOf(network).getStacks(), new Rs2ItemStack(ItemStubs.DIRT, 160), new Rs2ItemStack(ItemStubs.GLASS, 140));
 
         assertThat(diskDrive.getStored()).isEqualTo(150 + 10 + 140);
     }
@@ -244,98 +285,40 @@ class DiskDriveNetworkNodeTest {
     void Test_extracting() {
         // Arrange
         StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
-        StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
-        StorageDisk<Rs2ItemStack> storageDisk3 = new ItemDiskStorage(100);
-
         storageDisk1.insert(new Rs2ItemStack(ItemStubs.DIRT), 50, Action.EXECUTE);
         storageDisk1.insert(new Rs2ItemStack(ItemStubs.GLASS), 50, Action.EXECUTE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk1);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk1);
 
+        StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
         storageDisk2.insert(new Rs2ItemStack(ItemStubs.DIRT), 50, Action.EXECUTE);
         storageDisk2.insert(new Rs2ItemStack(ItemStubs.GLASS), 50, Action.EXECUTE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk2);
+        storageDiskProviderManager.setDiskInSlot(2, storageDisk2);
 
+        StorageDisk<Rs2ItemStack> storageDisk3 = new ItemDiskStorage(100);
         storageDisk3.insert(new Rs2ItemStack(ItemStubs.SPONGE), 10, Action.EXECUTE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(3, storageDisk3);
+        storageDiskProviderManager.setDiskInSlot(3, storageDisk3);
 
-        // Act & assert
-        Optional<Rs2ItemStack> stack1 = diskDrive.extract(new Rs2ItemStack(ItemStubs.DIRT), 85, Action.EXECUTE);
+        diskDrive.initialize(storageDiskProviderManager);
+        itemStorageOf(network).invalidate();
 
-        assertThat(stack1).isPresent();
-        assertItemStack(stack1.get(), new Rs2ItemStack(ItemStubs.DIRT, 85));
+        // Act
+        Optional<Rs2ItemStack> extracted = itemStorageChannelOf(network).extract(new Rs2ItemStack(ItemStubs.DIRT), 85, Action.EXECUTE);
+
+        // Assert
+        assertThat(extracted).isPresent();
+        assertItemStack(extracted.get(), new Rs2ItemStack(ItemStubs.DIRT, 85));
 
         assertItemStackListContents(storageDisk1.getStacks(), new Rs2ItemStack(ItemStubs.GLASS, 50));
         assertItemStackListContents(storageDisk2.getStacks(), new Rs2ItemStack(ItemStubs.GLASS, 50), new Rs2ItemStack(ItemStubs.DIRT, 15));
         assertItemStackListContents(storageDisk3.getStacks(), new Rs2ItemStack(ItemStubs.SPONGE, 10));
+        assertItemStackListContents(
+                itemStorageChannelOf(network).getStacks(),
+                new Rs2ItemStack(ItemStubs.GLASS, 100),
+                new Rs2ItemStack(ItemStubs.DIRT, 15),
+                new Rs2ItemStack(ItemStubs.SPONGE, 10)
+        );
 
         assertThat(diskDrive.getStored()).isEqualTo(125);
-    }
-
-    @Test
-    void Test_getting_stacks() {
-        // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
-        StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
-        StorageDisk<Rs2ItemStack> storageDisk3 = new ItemDiskStorage(100);
-
-        storageDisk1.insert(new Rs2ItemStack(ItemStubs.DIRT), 50, Action.EXECUTE);
-        storageDisk1.insert(new Rs2ItemStack(ItemStubs.GLASS), 50, Action.EXECUTE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk1);
-
-        storageDisk2.insert(new Rs2ItemStack(ItemStubs.DIRT), 50, Action.EXECUTE);
-        storageDisk2.insert(new Rs2ItemStack(ItemStubs.GLASS), 50, Action.EXECUTE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk2);
-
-        storageDisk3.insert(new Rs2ItemStack(ItemStubs.SPONGE), 10, Action.EXECUTE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(3, storageDisk3);
-
-        // Act
-        Collection<Rs2ItemStack> stacks = diskDrive.getStacks();
-
-        // Assert
-        assertItemStackListContents(stacks, new Rs2ItemStack(ItemStubs.DIRT, 100), new Rs2ItemStack(ItemStubs.GLASS, 100), new Rs2ItemStack(ItemStubs.SPONGE, 10));
-        assertThat(diskDrive.getStored()).isEqualTo(210);
-    }
-
-    @Test
-    void Test_storage_should_change_when_removing_a_disk() {
-        // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk1);
-
-        StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk2);
-
-        // Act
-        Optional<Rs2ItemStack> remainderBeforeRemovingDisk = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 105, Action.SIMULATE);
-        diskDrive.getFakeStorageDiskProviderManager().removeDiskInSlot(2);
-        Optional<Rs2ItemStack> remainderAfterRemovingDisk = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 105, Action.EXECUTE);
-
-        // Assert
-        assertThat(remainderBeforeRemovingDisk).isEmpty();
-
-        assertThat(remainderAfterRemovingDisk).isPresent();
-        assertItemStack(remainderAfterRemovingDisk.get(), new Rs2ItemStack(ItemStubs.DIRT, 5));
-    }
-
-    @Test
-    void Test_storage_should_change_when_inserting_a_disk() {
-        // Arrange
-        StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk1);
-
-        StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
-
-        // Act
-        Optional<Rs2ItemStack> remainderBeforeInsertingDisk = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 105, Action.SIMULATE);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, storageDisk2);
-        Optional<Rs2ItemStack> remainderAfterInsertingDisk = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 105, Action.EXECUTE);
-
-        // Assert
-        assertThat(remainderBeforeInsertingDisk).isPresent();
-        assertItemStack(remainderBeforeInsertingDisk.get(), new Rs2ItemStack(ItemStubs.DIRT, 5));
-
-        assertThat(remainderAfterInsertingDisk).isEmpty();
     }
 
     @Test
@@ -346,7 +329,9 @@ class DiskDriveNetworkNodeTest {
         diskDrive.setFilterTemplates(Arrays.asList(new Rs2ItemStack(ItemStubs.GLASS), new Rs2ItemStack(ItemStubs.STONE)));
 
         StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Act
         Rs2ItemStack glassWithTag = new Rs2ItemStack(ItemStubs.GLASS, 12, "myTag");
@@ -372,21 +357,20 @@ class DiskDriveNetworkNodeTest {
         diskDrive.setAccessMode(accessMode);
 
         StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Act
         Optional<Rs2ItemStack> remainder = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 5, Action.EXECUTE);
 
         // Assert
         switch (accessMode) {
-            case INSERT_EXTRACT:
-            case INSERT:
-                assertThat(remainder).isEmpty();
-                break;
-            case EXTRACT:
+            case INSERT_EXTRACT, INSERT -> assertThat(remainder).isEmpty();
+            case EXTRACT -> {
                 assertThat(remainder).isPresent();
                 assertItemStack(remainder.get(), new Rs2ItemStack(ItemStubs.DIRT, 5));
-                break;
+            }
         }
     }
 
@@ -397,7 +381,9 @@ class DiskDriveNetworkNodeTest {
         diskDrive.setAccessMode(accessMode);
 
         StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+
+        diskDrive.initialize(storageDiskProviderManager);
 
         storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 20, Action.EXECUTE);
 
@@ -406,14 +392,11 @@ class DiskDriveNetworkNodeTest {
 
         // Assert
         switch (accessMode) {
-            case INSERT_EXTRACT:
-            case EXTRACT:
+            case INSERT_EXTRACT, EXTRACT -> {
                 assertThat(extracted).isPresent();
                 assertItemStack(extracted.get(), new Rs2ItemStack(ItemStubs.DIRT, 5));
-                break;
-            case INSERT:
-                assertThat(extracted).isEmpty();
-                break;
+            }
+            case INSERT -> assertThat(extracted).isEmpty();
         }
     }
 
@@ -421,9 +404,10 @@ class DiskDriveNetworkNodeTest {
     void Test_inserting_when_inactive() {
         // Arrange
         diskDrive.setRedstoneMode(RedstoneMode.HIGH);
+        diskDrive.initialize(storageDiskProviderManager);
 
         StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
 
         // Act
         Optional<Rs2ItemStack> remainder = diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 5, Action.EXECUTE);
@@ -439,7 +423,9 @@ class DiskDriveNetworkNodeTest {
         diskDrive.setRedstoneMode(RedstoneMode.HIGH);
 
         StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+
+        diskDrive.initialize(storageDiskProviderManager);
 
         storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 20, Action.EXECUTE);
 
@@ -451,59 +437,104 @@ class DiskDriveNetworkNodeTest {
     }
 
     @Test
-    void Test_getting_stacks_when_inactive() {
+    void Test_disk_state_change_listener_should_not_be_called_when_not_necessary_on_extracting() {
         // Arrange
         StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, storageDisk);
-
-        diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 20, Action.EXECUTE);
+        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 76, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Act
-        Collection<Rs2ItemStack> stacksBeforeInactive = diskDrive.getStacks();
-
-        diskDrive.setRedstoneMode(RedstoneMode.HIGH);
-
-        Collection<Rs2ItemStack> stacksAfterInactive = diskDrive.getStacks();
+        itemStorageChannelOf(network).extract(new Rs2ItemStack(ItemStubs.DIRT), 1, Action.EXECUTE);
 
         // Assert
-        assertThat(stacksBeforeInactive).isNotEmpty();
-        assertThat(stacksAfterInactive).isEmpty();
+        verify(diskDriveListener, never()).onDiskChanged();
     }
 
     @Test
-    void Test_changing_to_inactive_should_omit_items_from_storage_channel() {
+    void Test_disk_state_change_listener_should_not_be_called_when_not_necessary_on_inserting() {
         // Arrange
-        StorageDisk<Rs2ItemStack> disk = new ItemDiskStorage(10);
-        disk.insert(new Rs2ItemStack(ItemStubs.DIRT), 10, Action.EXECUTE);
-
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(0, disk);
-
-        network.getComponent(ItemStorageNetworkComponent.class).invalidate();
-
-        diskDrive.setRedstoneMode(RedstoneMode.HIGH);
+        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Act
-        Collection<Rs2ItemStack> stacksBeforeStateChange = network.getComponent(ItemStorageNetworkComponent.class).getStorageChannel().getStacks();
-        diskDrive.onActiveChanged(false);
-        Collection<Rs2ItemStack> stacksAfterStateChange = network.getComponent(ItemStorageNetworkComponent.class).getStorageChannel().getStacks();
+        itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.DIRT), 74, Action.EXECUTE);
 
         // Assert
-        assertThat(stacksBeforeStateChange).isNotEmpty();
-        assertThat(stacksAfterStateChange).isEmpty();
+        verify(diskDriveListener, never()).onDiskChanged();
     }
 
-    @Test
-    void Test_calculating_energy_usage() {
+    @ParameterizedTest
+    @EnumSource(Action.class)
+    void Test_disk_state_change_listener_should_be_called_when_necessary_on_extracting(Action action) {
         // Arrange
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(0, new ItemDiskStorage(1));
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(1, new ItemDiskStorage(1));
-        diskDrive.getFakeStorageDiskProviderManager().removeDiskInSlot(1);
-        diskDrive.getFakeStorageDiskProviderManager().setDisk(2, new ItemDiskStorage(1));
+        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
+        storageDisk.insert(new Rs2ItemStack(ItemStubs.DIRT), 75, Action.EXECUTE);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+        diskDrive.initialize(storageDiskProviderManager);
 
         // Act
-        long energyUsage = diskDrive.getEnergyUsage();
+        itemStorageChannelOf(network).extract(new Rs2ItemStack(ItemStubs.DIRT), 1, action);
+        itemStorageChannelOf(network).extract(new Rs2ItemStack(ItemStubs.DIRT), 1, action);
 
         // Assert
-        assertThat(energyUsage).isEqualTo(BASE_USAGE + (USAGE_PER_DISK * 2));
+        VerificationMode expectedTimes = action == Action.EXECUTE ? times(1) : never();
+
+        verify(diskDriveListener, expectedTimes).onDiskChanged();
+    }
+
+    @ParameterizedTest
+    @EnumSource(Action.class)
+    void Test_disk_state_change_listener_should_be_called_when_necessary_on_inserting(Action action) {
+        // Arrange
+        StorageDisk<Rs2ItemStack> storageDisk = new ItemDiskStorage(100);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk);
+        diskDrive.initialize(storageDiskProviderManager);
+        diskDrive.insert(new Rs2ItemStack(ItemStubs.DIRT), 74, Action.EXECUTE);
+
+        // Act
+        itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.DIRT), 1, action);
+        itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.DIRT), 1, action);
+
+        // Assert
+        VerificationMode expectedTimes = action == Action.EXECUTE ? times(1) : never();
+
+        verify(diskDriveListener, expectedTimes).onDiskChanged();
+    }
+
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    void Test_setting_priority(boolean oneHasPriority) {
+        // Arrange
+        StorageDisk<Rs2ItemStack> storageDisk1 = new ItemDiskStorage(100);
+        storageDiskProviderManager.setDiskInSlot(1, storageDisk1);
+        diskDrive.initialize(storageDiskProviderManager);
+
+        StorageDisk<Rs2ItemStack> storageDisk2 = new ItemDiskStorage(100);
+        FakeStorageDiskProviderManager storageDiskProviderManager2 = new FakeStorageDiskProviderManager();
+        storageDiskProviderManager2.setDiskInSlot(1, storageDisk2);
+        FakeNetworkNodeContainer<DiskDriveNetworkNode> diskDrive2 = createDiskDriveContainer(network, storageDiskProviderManager2, mock(DiskDriveListener.class));
+        diskDrive2.getNode().initialize(storageDiskProviderManager2);
+
+        if (oneHasPriority) {
+            diskDrive.setPriority(5);
+            diskDrive2.getNode().setPriority(2);
+        } else {
+            diskDrive.setPriority(2);
+            diskDrive2.getNode().setPriority(5);
+        }
+
+        // Act
+        itemStorageChannelOf(network).insert(new Rs2ItemStack(ItemStubs.DIRT), 1, Action.EXECUTE);
+
+        // Assert
+        if (oneHasPriority) {
+            assertThat(diskDrive.getStacks()).isNotEmpty();
+            assertThat(diskDrive2.getNode().getStacks()).isEmpty();
+        } else {
+            assertThat(diskDrive.getStacks()).isEmpty();
+            assertThat(diskDrive2.getNode().getStacks()).isNotEmpty();
+        }
     }
 }