+++ CHANGELOG.md
@@ -12,11 +12,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Fix crash when transferring items in the Controller screen
+- Fix Disk Drive leds not being stable
+- Fix block variants not being present on Patchouli book entries
 
 ### Changed
 
 - Ported to Minecraft 1.17.1
-- Add block variants to Patchouli book entries
 - Implemented a new networking system
 
 ## [2.0.0-milestone.1.0] - 2021-05-21